Added raw method to execute sql queries.

+++ gcov.sh
@@ -2,9 +2,16 @@
 FOLDER=gcov
 export BOOST_TEST_SHOW_PROGRESS="yes"
 
+a=$(which lcov)
+if [ -z $a ]
+then
+    echo "Couldn't find lcov"
+    exit 1
+fi
+
 rm -rf gcov
 cmake . -DCOVERAGE=1 && \
-make && \
+make -j3 && \
 mkdir -p "$FOLDER" && \
 ./unittest && \
 lcov --checksum -o "$FOLDER/gcov.info" --no-external --capture -d CMakeFiles -d torrentsync -d test &>/dev/null
